reorganize btc spv. drop pointless conversion funcs
